+++ .github/workflows/telegram-bot.yml
@@ -23,7 +23,6 @@ jobs:
       PR_MERGED: ${{ github.event.pull_request.merged_by.login }}
       REVIEW_STATE: ${{ github.event.review.state }}
       REVIEW_AUTHOR: ${{ github.event.review.user.login }}
-      REVIEW_COMMENT: ${{ github.event.review.body }}
       COMMENT_AUTHOR: ${{ github.event.comment.user.login }}
       COMMENT_URL: ${{ github.event.comment.html_url }}
       COMMENT_BODY: ${{ github.event.comment.body }}
@@ -52,10 +51,7 @@ jobs:
 
         elif [[ ! -z "${{ github.event.review }}" && "${{ env.REVIEW_STATE }}" == "changes_requested" ]]; then
           TEXT=`echo -e "<b>${{ env.REVIEW_AUTHOR }}</b> requested changes for <a href=\"${{ env.PR_URL }}\">PR#${{ env.PR_NUMBER }}</a>"`
-        elif [[ ! -z "${{ github.event.review }}" && "${{ env.REVIEW_STATE }}" == "commented" ]]; then
-          ESCAPED_TEXT=`echo -e "${{ env.REVIEW_COMMENT }}"| sed 's/\&/\&amp;/g' | sed 's/</\&lt;/g' | sed 's/>/\&gt;/g'`
-          TEXT=`echo -e "<b>${{ env.REVIEW_AUTHOR }}</b> posted the following comment to <a href=\"${{ env.PR_URL }}\">PR#${{ env.PR_NUMBER }}</a>:\n<i>$ESCAPED_TEXT</i>"`
-        elif [[ ! -z "${{ github.event.review }}" ]]; then
+        elif [[ ! -z "${{ github.event.review }}" && "${{ env.REVIEW_STATE }}" != "changes_requested" ]]; then
           TEXT=`echo -e "<b>${{ env.REVIEW_AUTHOR }}</b> ${{ env.REVIEW_STATE }} <a href=\"${{ env.PR_URL }}\">PR#${{ env.PR_NUMBER }}</a>"`
         elif [[ -z "${{ github.event.review }}" && "${{ github.event.action }}" == "submitted" ]]; then
           TEXT=`echo -e "Due to GitHub Actions bug we cannot identify, who approved <a href=\"${{ env.PR_URL }}\">PR#${{ env.PR_NUMBER }}</a>"`
@@ -67,7 +63,7 @@ jobs:
         fi
 
         if [[ ! -z $TEXT ]]; then
-          curl --get --data-urlencode "chat_id=${{ secrets.TELEGRAM_CHAT_ID }}" --data-urlencode "disable_web_page_preview=True" \
+          curl --get --data-urlencode "chat_id=${{ secrets.TELEGRAM_CHAT_ID }}" \
           --data-urlencode "text=$TEXT" --data-urlencode "parse_mode=HTML" $URL
         fi
