+++ CHANGELOG.rst
@@ -1,23 +1,6 @@
 Change log
 ==========
 
-0.19.15 (2023-05-10)
---------------------
-
-New features
-~~~~~~~~~~~~
-
-* **analyzer:** code report based on clang-format
-
-Bug fixes
-~~~~~~~~~
-
-* **config:** fixed returning "None" instead of "False" for Step custom keys set to "False"
-* **artifact:** remove redundant Static_analysis_report.json from artifacts
-* **analyzer:** incorrect program name in help
-* **report:** set exit code even if reporting crashes
-
-
 0.19.14 (2022-11-14)
 --------------------
