+++ doc/code_report.rst
@@ -6,7 +6,6 @@ The following analysing modules (analysers) are currently added to Universum:
    * `pylint`_
    * `mypy`_
    * `uncrustify`_
-   * `clang-format`_
 
 Analysers are separate scripts, fully compatible with Universum. It is possible to use them
 as independent Python modules.
@@ -63,6 +62,7 @@ Pylint
 
 .. argparse::
     :ref: universum.analyzers.pylint.pylint_argument_parser
+    :prog: {python} -m universum.analyzers.pylint
 
 Config example for ``universum.analyzers.pylint``:
 
@@ -99,6 +99,7 @@ Mypy
 
 .. argparse::
     :ref: universum.analyzers.mypy.mypy_argument_parser
+    :prog: {python} -m universum.analyzers.mypy
 
 Config example for ``universum.analyzers.mypy``:
 
@@ -135,6 +136,7 @@ Uncrustify
 
 .. argparse::
     :ref: universum.analyzers.uncrustify.uncrustify_argument_parser
+    :prog: {python} -m universum.analyzers.uncrustify
     :nodefault:
 
 Config example for ``universum.analyzers.uncrustify``:
@@ -144,7 +146,7 @@ Config example for ``universum.analyzers.uncrustify``:
     from universum.configuration_support import Configuration, Step
 
     configs = Configuration([Step(name="uncrustify", code_report=True, command=[
-        "{python}", "-m", "universum.analyzers.uncrustify",  "--files", "/home/user/workspace/temp/*.c",
+        "{python}", "-m", "universum.analyzers.uncrustify",  "--files", "/home/user/workspace/temp",
         "--cfg-file", "file_name.cfg", "--result-file", "${CODE_REPORT_FILE}", "--output-directory", "uncrustify"
     ])])
 
@@ -162,38 +164,4 @@ will produce this list of configurations:
 .. testoutput::
 
     $ ./.universum.py
-    [{'name': 'uncrustify', 'code_report': True, 'command': '{python} -m universum.analyzers.uncrustify --files /home/user/workspace/temp/*.c --cfg-file file_name.cfg --result-file ${CODE_REPORT_FILE} --output-directory uncrustify'}]
-
-Clang-format
-------------
-
-.. argparse::
-    :ref: universum.analyzers.clang_format.clang_format_argument_parser
-    :nodefault:
-
-Config example for ``universum.analyzers.clang_format``:
-
-.. testcode::
-
-    from universum.configuration_support import Configuration, Step
-
-    configs = Configuration([Step(name="clang-format", code_report=True, command=[
-        "{python}", "-m", "universum.analyzers.clang-format",  "--files", "/home/user/workspace/temp/*.c",
-        "--result-file", "${CODE_REPORT_FILE}", "--output-directory", "clang-format", "-e", "clang-format-15",
-    ])])
-
-    if __name__ == '__main__':
-        print(configs.dump())
-
-will produce this list of configurations:
-
-.. testcode::
-    :hide:
-
-    print("$ ./.universum.py")
-    print(configs.dump())
-
-.. testoutput::
-
-    $ ./.universum.py
-    [{'name': 'clang-format', 'code_report': True, 'command': '{python} -m universum.analyzers.clang-format --files /home/user/workspace/temp/*.c --result-file ${CODE_REPORT_FILE} --output-directory clang-format -e clang-format-15'}]
+    [{'name': 'uncrustify', 'code_report': True, 'command': '{python} -m universum.analyzers.uncrustify --files /home/user/workspace/temp --cfg-file file_name.cfg --result-file ${CODE_REPORT_FILE} --output-directory uncrustify'}]
