+++ doc/conf.py
@@ -50,7 +50,7 @@ copyright = u'Samsung Electronics Co., Ltd.'
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = 'en'
+language = None
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
