+++ tests/deployment_utils.py
@@ -42,8 +42,7 @@ class ExecutionEnvironment:
         if self._container:
             self.request.raiseerror("Container is already running, no dirs can be bound!")
         for directory in directories:
-            absolute_dir = str(pathlib.Path(directory).absolute())
-            self._volumes[absolute_dir] = {'bind': absolute_dir, 'mode': 'rw'}
+            self._volumes[directory] = {'bind': directory, 'mode': 'rw'}
 
     def add_environment_variables(self, variables):
         if self._container:
